Backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
mysql-connector-python==8.0.33
prometheus-fastapi-instrumentator